Detect Windows Console Improvements and Change Shell Accordingly
This is another take on meteor/meteor#8082 which instead of using the `process.title` (which I'm thinking might break on internationalized versions of Windows?) uses the Windows Console SDK GetConsoleMode output mode.

The downside to this approach is that there is an additional half-second delay during the initial launch while the PowerShell script executes.

This would be an alternative fix for meteor/meteor#6785 but can probably wait until this actually proves to be a problem (my hunch about the title might be wrong?)