[BISECT] Allow building the dev bundle from Node.js source.
This introduces two *additional* environment variables,
`NODE_COMMIT_HASH` and/or `NODE_FROM_SRC`, which can be set
either in the environment, or preferably from
`build-dev-bundle-common.sh`.

`NODE_FROM_SRC` is implied by `NODE_COMMIT_HASH`, but in either case, it
can also be set explicitly to `debug` to build with the `--debug` flag.

This was incredibly helpful for `git bisect`-ing a Node.js issue.